[new feature] New reborn system
Before you need to press the reborn button to active the
reborn ability and die within 10 seconds, you will be
able to reborn. For the new version, you don't need to
active the reborn button. When you die and if you have
reborn in your ability, button will show up and ask you
to reborn.